Refactor SearchPanelControls to drop navigation into its own component
